Refactor #68 - rm selection internal package
